rgr/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
cryptography==40.0.2
passlib[argon2]==1.7.4
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.15